Correct .gitignore for src folder
We need to explicitly allow the src folder as well as its contents to
undo the * rule.  I have not checked whether this is Git version
dependent, but it seems reasonable that this change will not break
other Git versions as it can, at worst, only allow more desired
contents to sip through.