Support for RDFa in blocks which have no child nodes
It's unclear what the exact semantics are for a block of content to be
considered a block in html.  We're not sure if there are ways in which
a text node could be a block.  We should figure that out to be sure.
Furthermore, we're not sure the implementation of "tag" is done
correctly in the rdfa-walker where it assumes one needs to have
children in order to be a block.  It might be that that is the thing
tag wants to express, but it's not exactly something you'd assume from
the name.  The combination of these two inonsistencies made it so that
we got into undesired behaviour here.  Where we assumed all non-tag
nodes had to be text or comment nodes (but certainly not nodes with
RDFa attached to them).  As this turns out to be incorrect, we've
changed the implementation in the least impactful way we could.  This
creates somewhat overly convoluted code hence it's suboptimal, but it
gives us the highest odds at landing on a fully working
implementation.

If you're looking at this commit and you're wondering if this could be
done in a simpler way: yes!  It probably can.  Do test it both in the
browser as well as in jsdom as both may use marawa.