Support running regular tests as well as debugging tests
When debugging and writing tests it's nice if the tests are being
watched and if you can attach the node inspector to the test suite.
Running `npm run debug-test` will watch the files for changes and will
break in a debugger session.  This allows you to hook into the
debugger from chrome and use it's devtools to inspect why certain
tests fail or succeed.