Treat regular dom nodes as they are instead of requiring children
The node walker incorrectly labeled nodes which didn't have children
as a node which isn't a tag.  Cleaning up this code as that's most
likely not what was intended.  Other code depending on this unexpected
behaviour should be updated.

We assume the misconception as created by working out a solution and
not coupling it back to the browser specification.  Our current
research shows cleaner approaches exist.