Yield full logicalblock as minimal node for RDFa ranges
When requesting the context scanner for a range, we should yield
meaningful content.  The new approach is to consider all lowest-level
logicalBlocks as a meaningfull set of content.  This alleviates issues
with accidentally receiving too little content because a seemingly
unimportant tag was inserted.  Logical blocks try to reason on the
level of content which humans likely consider to be one element to
reason on (eg: a name, a title, a paragraph, ...).