Improved URI resolution: extract and improve prefixed URI resolution
Extracted the function to resolve prefixed URI's

The removed part of the code contained the old procedure.
The updated procedure is in the new function.

The improvement being that if no default vocab is provided, or prefix is unknown,
the URI is returned as such. Previously it returned `undefined:uri`.

This leaves standard prefixes like mailto and
data URI's (e.g base64 encoded images) untouched.